Initial Commit
In this project I am working on a dataset on a survey taken by employees in data analytics/science. After cleaning the data, I created a dashboard to show the levels of salary across different roles, how happy respondents were, how difficult it was for them to land a job in data, the countries of them etc.